third_party/blink/web_tests/platform/linux-chrome/external/wpt/html/cross-origin-opener-policy/reporting/access-reporting/access-from-coop-page-to-opener_coop-ro_cross-origin.https-expected.txt

This is a testharness.js-based test.
[FAIL] access-from-coop-page-to-opener, cross-origin
  assert_equals: expected (string) "coop" but got (undefined) undefined
[FAIL] access-from-coop-page-to-opener, cross-origin + redirect
  assert_equals: expected (string) "coop" but got (undefined) undefined
Harness: the test ran to completion.
